Move code from VintageNet to VintageNetWizard
This moves the code out of the VintageNet namespace and into its own
thing.
