+++ lib/vintage_net_wizard.ex
@@ -1,0 +1,5 @@
+defmodule VintageNetWizard do
+  @moduledoc """
+  Documentation for VintageNetWizard.
+  """
+end
