--- lib/wizard.ex
@@ -1,2 +1,0 @@
-defmodule VintageNet.Wizard do
-end
